Fix ResultSet.getObject(*, BigInteger.class) for DECIMAL values

+++ h2/src/main/org/h2/jdbc/JdbcResultSet.java
@@ -3773,7 +3773,7 @@ public class JdbcResultSet extends TraceObject implements ResultSet, JdbcResultS
         if (type == BigDecimal.class) {
             return type.cast(value.getBigDecimal());
         } else if (type == BigInteger.class) {
-            return type.cast(BigInteger.valueOf(value.getLong()));
+            return type.cast(value.getBigDecimal().toBigInteger());
         } else if (type == String.class) {
             return type.cast(value.getString());
         } else if (type == Boolean.class) {
